modified database connection handling during queued jobs
